+++ Utility.Test/Extension/EnumerableExtensionTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
 using Messerli.Utility.Extension;
 using Xunit;
 
@@ -18,12 +19,27 @@ namespace Messerli.Utility.Test.Extension
             Assert.True(elements.All(element => element.SideEffect));
         }
 
+        [Fact]
+        public async Task ForEachExecutesActionAsync()
+        {
+            var elements = System.Linq.Enumerable.Repeat(new ForEachExecutor(), 17);
+
+            // ReSharper disable once PossibleMultipleEnumeration
+            await elements.ForEachAsync(element => element.ExecuteAsync());
+
+            // ReSharper disable once PossibleMultipleEnumeration
+            Assert.True(elements.All(element => element.SideEffect));
+        }
+
         private class ForEachExecutor
         {
             public bool SideEffect { get; private set; }
 
             public void Execute()
                 => SideEffect = true;
+
+            public Task ExecuteAsync()
+                => Task.Run(Execute);
         }
     }
 }